Store and send keys as BaseRef objects
Keys now have a secret component which prevents the client from
simply enumerating the contents of the server.

It's the first six bytes of a SHA1 hash of (a secret and) the key.
which is enough to make a random guess sufficiently unlikely.

TODO: block the client when that happens.