Reorganized passing the decoder's loader, SQL sessions
The symptom was that pypy was unable to encode a SQL object because the
session was garbage-collected and freed too early, which marked the objects
that had to be encoded as invalid.

The solution: use Werkzeug's thread-local objects, and add a wrapper to
thread startup so that these _will_ be deallocated when a thread ends.

… and if we have these anyway, add a localstack for decoding.

To make sure that deallocation is always performed, add wrappers to local
objects and thread creation. The latter might allow us to use OS threads.
Maybe. TODO, if I want that: Add a whole bunch of locks, as neither the
client cache nor some transports are thread safe at the moment.