Fix encoding of multiply-referenced objects within other objects
Do this by pre-loading the cache and replacing the original occurrence of
the object with a reference.
Do this for all objects which are not used recursively, i.e. whose
declaration is finished before they're referenced.
Seed the cache in object order, i.e. the object which is completed first
gets added to the cache first.