Sync client and server
The problem: Client sends a request, sever sends a bunch of broadcasts and
a reply, client processes the reply without waiting for the broadcasts --
which may trigger processing required for correct further processing.

Solve this by adding an incrementing message ID to the server broadcasts
and mention that ID in the reply.

Problem: Doing this is not enough, we need to delay unpacking the reply
until broadcasts have been processed.

Solve this by splitting up message decoding.